Login stuff The axolotl (/ˈæksəlɒtəl/; from Classical Nahuatl: āxōlōtl [aːˈʃoːloːtɬ] (listen)), Ambystoma mexicanum,[3] is a paedomorphic salamander related to the tiger salamander.[3][4][5] Axolotls are unusual among amphibians in that they reach adulthood without undergoing metamorphosis. Instead of taking to the land, adults remain aquatic and gilled. The species was originally found in several lakes underlying what is now Mexico City, such as Lake Xochimilco and Lake Chalco.[1] These lakes were drained by Spanish settlers after the conquest of the Aztec Empire, leading to the destruction of much of the axolotl’s natural habitat.
Axolotls should not be confused with the larval stage of the closely related tiger salamander (A. tigrinum), which are widespread in much of North America and occasionally become paedomorphic. Neither should they be confused with mudpuppies (Necturus spp.), fully aquatic salamanders from a different family that are not closely related to the axolotl but bear a superficial resemblance.[6]

As of 2020, wild axolotls were near extinction[7][8] due to urbanization in Mexico City and consequent water pollution, as well as the introduction of invasive species such as tilapia and perch. They are listed as critically endangered in the wild, with a decreasing population of around 50 to 1000 adult individuals, by the International Union for Conservation of Nature and Natural Resources (IUCN) and are listed under Appendix II of the Convention on International Trade in Endangered Species (CITES).[2] Axolotls are used extensively in scientific research due to their ability to regenerate limbs, gills and parts of their eyes and brains.[9] Axolotls were also sold as food in Mexican markets and were a staple in the Aztec diet.[10]