The Sun is, by far, the nearest star to Earth. The distance from Earth to Sun is 150 million kilometers (km), which equals 1 astronomical unit (AU), which equals 8.3 light minutes (that is, the distance traveled by light in 8.3 minutes). By contrast, the distance from Earth to Proxima Centauri (the next nearest star) is 40 trillion kilometers, which equals 270,000 astronomical units, which equals 4.2 light years (ly). It's much easier to study the nearby Sun than the much-more-distant other stars. The Sun is made of dense gas. The average density of the Sun is about 1400 kilograms per cubic meter, 40 percent greater than the density of water. The Sun remains gaseous (rather than liquid or solid) at these high densities because of its extraordinarily high temperature, which prevents the ions within the Sun from bonding together. The gas of which the Sun is made is mostly ionized hydrogen and ionized helium, the two simplest chemical elements. The Sun thus contains positively charged hydrogen nuclei, positively charged helium nuclei, and negatively charged electrons. Because photons scatter readily from free electrons, the Sun is opaque (you can't see through the Sun to see the stars behind it).
As you know from Astronomy 161, a hot, dense, opaque object (such as the Sun, or the filament of a light bulb) produces a continuous blackbody spectrum of light. The temperature at the Sun's surface is 5800 Kelvin; the Sun's blackbody spectrum thus peaks at a wavelength of 500 nanometers, in the visible range of the spectrum. [For a review of blackbody radiation, see chapter 5 of the textbook, or the relevant Astronomy 161 web page.] The luminosity of the Sun (that is, the amount of energy it radiates away per second in the form of light) is very large, thanks to the Sun's high temperature and large surface area. The total luminosity of the Sun is 3.9 x 1026 watts. (That is, 390 trillion trillion watts.)

(2) The Sun remains hot because it is powered by nuclear fusion in its core.
The Sun has been shining away for 4.6 billion years. For well over 3 billion years, according to the fossil record, life has existed on Earth. The fact that life was not scorched or frozen out of existence during that time span indicates that the luminosity of the Sun has been fairly steady. What energy source has kept the Sun hot for billions of years, instead of cooling off like an unplugged iron?
What about chemical sources, such as the burning of hydrogen? (We know there's plenty of hydrogen in the Sun.) Burning one kilogram of hydrogen to form water, by the reaction
2H + O -> H2O + energy,
releases 140 million joules of energy. (The `joule' is the standard unit of energy in the metric system. It is the energy required to keep a 1 watt light bulb lit up for 1 second. 140 million joules is equivalent to 33,000 Calories or 40 kilowatt-hours.) If the Sun consisted entirely of hydrogen (which it doesn't), and if you could find a source of oxygen with which to burn it (never mind where all that oxygen would come from), it could keep the Sun at its present luminosity for a mere 20,000 years.

A more potent energy source is needed to keep the Sun shining for billions of years. What about nuclear fusion? Four hydrogen nuclei can be fused together into a helium nucleus, accompanied by the release of energy:
4H -> He + energy.
Fusing one kilogram of hydrogen into helium releases 630 trillion joules of energy. That's nearly 5 million times what you'd get by burning the same hydrogen to get water. Thus, fusion of hydrogen will keep the Sun shining at a constant luminosity not for 20 thousand years, but for as much as 100 billion years.

(3) When hydrogen is fused into helium, matter is converted into energy.
The Sun, like most stars, is powered by the fusion of hydrogen into helium (what the textbook misleadingly calls `hydrogen burning'). You may well ask, ``Where does all that energy come from when hydrogen is fused into helium? Isn't energy conserved?'' There is another question you can ask. The total mass of four hydrogen nuclei is 6.693 x 10-27 kilograms. The mass of a helium nucleus is 6.645 x 10-27 kilograms. Thus, there is a loss of 0.048 x 10-27 kilograms every time a helium nucleus is formed by fusion. The loss is small, but real. You may then ask, ``Where does that mass go when hydrogen is fused into helium? Isn't mass conserved?''
Albert Einstein provided us with an answer to our questions. He pointed out that mass can be converted into energy, and vice versa. The `exchange rate' between mass and energy is given by the famous equation
E = m c2
where E = energy, m = mass, and c = the speed of light in a vacuum (300,000,000 ometers/second). Thus, the total amount of MASS+ENERGY is conserved.

Every second, as we have seen, the Sun creates E = 3.9 x 1026 joules of energy. To balance the books, every second the Sun destroys m = E/c2 = 4.3 x 109 kilograms of mass. This mass loss, equivalent to more than 4 million tons per second, is accomplished by fusing 600 million tons of hydrogen into 596 million tons of helium.

It should be noted that fusion of atomic nuclei is difficult to initiate. Atomic nuclei are positively charged; thus each nucleus repels each other nucleus. In order to fuse two nuclei together, they must be brought very close together, so that the strong (but short-range) nuclear force that holds nuclei together can do its work. In order for two positively charged nuclei to overcome their electric repulsion and come very close, they have to collide at high speeds. High speeds for individual nuclei imply that the gas in which they exist is very hot (more than 10 million degrees Kelvin). These high temperatures are not found at the Sun's surface, but only in its hot, dense core.