Turning the other cheek From Wikipedia, the free encyclopedia Jump to navigationJump to search
Jesus taught turning the other cheek during the Sermon on the Mount.
Turning the other cheek is a phrase in Christian doctrine from the Sermon on the Mount that refers to responding to injury without revenge and allowing more injury. This passage is variously interpreted as accepting one's predicament, commanding nonresistance or advocating Christian pacifism.

Contents
1	Scriptural references
2	Interpretations
2.1	Christian anarchist interpretation
2.2	Nonviolent resistance interpretation
3	See also
4	References
5	Further reading
6	External links
Scriptural references
The phrase originates from the Sermon on the Mount in the New Testament. In the Gospel of Matthew chapter 5, an alternative for "an eye for an eye" is given by Jesus:

38You have heard that it was said, "An eye for an eye and a tooth for a tooth." 39But I say to you, Do not resist the one who is evil. But if anyone slaps you on the right cheek, turn to him the other also. 40And if anyone would sue you and take your tunic, let him have your cloak as well. 41And if anyone forces you to go one mile, go with him two miles. 42Give to the one who begs from you, and do not refuse the one who would borrow from you.

— Jesus Christ, English Standard Version (Matthew 5:38–42)
In the Sermon on the Plain[1] in the Gospel of Luke chapter 6, as part of his command to "love your enemies", Jesus says:

27But I say to you who hear, Love your enemies, do good to those who hate you, 28bless those who curse you, pray for those who abuse you. 29To one who strikes you on the cheek, offer the other also, and from one who takes away your cloak do not withhold your tunic either. 30Give to everyone who begs from you, and from one who takes away your goods do not demand them back. 31And as you wish that others would do to you, do so to them.

— Jesus Christ, English Standard Version (Luke 6:27–31)
Interpretations
This phrase, as with much of the Sermon on the Mount, has been subject to both literal and figurative interpretations.

Christian anarchist interpretation
Main article: Christian anarchism
According to this interpretation the passages call for total nonresistance to the point of facilitating aggression against oneself, and since human governments defend themselves by military force, some have advocated Christian anarchism, including Leo Tolstoy who elucidated his reasoning in his 1894 book The Kingdom of God Is Within You.

Nonviolent resistance interpretation
The scholar Walter Wink, in his book Engaging the Powers: Discernment and Resistance in a World of Domination, interprets the passage as ways to subvert the power structures of the time.[2]

At the time of Jesus, says Wink, striking backhand a person deemed to be of lower socioeconomic class was a means of asserting authority and dominance. If the persecuted person "turned the other cheek," the discipliner was faced with a dilemma: The left hand was used for unclean purposes, so a back-hand strike on the opposite cheek would not be performed. An alternative would be a slap with the open hand as a challenge or to punch the person, but this was seen as a statement of equality. Thus, by turning the other cheek, the persecuted was demanding equality.

Wink continues with an interpretation of handing over one's cloak in addition to one's tunic. The debtor has given the shirt off his back, a situation forbidden by Hebrew law as stated in Deuteronomy (24:10–13). By giving the lender the cloak as well, the debtor was reduced to nakedness. Wink notes that public nudity was viewed as bringing shame on the viewer, and not just the naked, as seen in Noah's case (Genesis 9:20–23).

Wink interprets the succeeding verse from the Sermon on the Mount as a method for making the oppressor break the law. The commonly invoked Roman law of Angaria allowed the Roman authorities to demand that inhabitants of occupied territories carry messages and equipment the distance of one mile post, but prohibited forcing an individual to go further than a single mile, at the risk of suffering disciplinary actions.[3] In this example, the nonviolent interpretation sees Jesus as placing criticism on an unjust and hated Roman law, as well as clarifying the teaching to extend beyond Jewish law.[4]